requirements.txt
aiohttp==3.10.10
azure-functions==1.21.3
fqdn==1.5.1
fuzzywuzzy==0.18.0
isoduration==20.11.0
jsonpointer==3.0.0
matplotlib==3.9.2
notebook==7.2.2
openpyxl==3.1.5
pandas==2.2.3
peopledatalabs==4.1.2
pipdeptree==2.23.4
pyproject-toml==0.1.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dotenv==1.0.1
tqdm==4.67.0
uri-template==1.3.0